+++ server/src/main/java/com/example/server/serviceLayer/FacadeObjects/PolicyFacade/SimpleDiscountFacade.java
@@ -2,18 +2,13 @@ package com.example.server.serviceLayer.FacadeObjects.PolicyFacade;
 
 import com.example.server.businessLayer.Market.Policies.DiscountPolicy.CompositeDiscount.MaxCompositeDiscount;
 import com.example.server.businessLayer.Market.Policies.DiscountPolicy.ConditionalDiscount;
-import com.example.server.businessLayer.Market.Policies.DiscountPolicy.DiscountState.AndCompositeDiscountLevelState;
 import com.example.server.businessLayer.Market.Policies.DiscountPolicy.DiscountState.DiscountLevelState;
 import com.example.server.businessLayer.Market.Policies.DiscountPolicy.DiscountType;
 import com.example.server.businessLayer.Market.Policies.DiscountPolicy.SimpleDiscount;
 import com.example.server.businessLayer.Market.ResourcesObjects.MarketException;
-import com.example.server.serviceLayer.FacadeObjects.AppointmentFacade;
-import com.example.server.serviceLayer.FacadeObjects.PolicyFacade.DiscountLevelStateFacade;
-import com.example.server.serviceLayer.FacadeObjects.PolicyFacade.DiscountTypeFacade;
-import com.example.server.serviceLayer.FacadeObjects.ShopOwnerAppointmentFacade;
 
 public class SimpleDiscountFacade extends DiscountTypeFacade {
-    public SimpleDiscountFacade(int percentageOfDiscount, DiscountLevelStateFacade discountLevelState) {
+    public SimpleDiscountFacade(double percentageOfDiscount, DiscountLevelStateFacade discountLevelState) {
         super (percentageOfDiscount, discountLevelState );
     }
